requirements.txt
fastapi
uvicorn
python-multipart
playwright
pillow
nest_asyncio
jinja2
pywin32; sys_platform == 'win32'
